# hdkeychain: Introduce v2 module.