--- a/docs/README.md
+++ b/docs/README.md
@@ -209,8 +209,9 @@ The following versioned modules are provided by dcrd repository:
   concurrency safe Decred network address manager
 * [connmgr](https://github.com/decred/dcrd/tree/master/connmgr) - Implements a
   generic Decred network connection manager
-* [hdkeychain](https://github.com/decred/dcrd/tree/master/hdkeychain) - Provides
-  an API for working with  Decred hierarchical deterministic extended keys
+* [hdkeychain/v2](https://github.com/decred/dcrd/tree/master/hdkeychain) -
+  Provides an API for working with  Decred hierarchical deterministic extended
+  keys
 * [gcs](https://github.com/decred/dcrd/tree/master/gcs) - Provides an API for
   building and using Golomb-coded set filters useful for light clients such as
   SPV wallets