--- a/hdkeychain/bench_test.go
+++ b/hdkeychain/bench_test.go
@@ -3,12 +3,10 @@
 // Use of this source code is governed by an ISC
 // license that can be found in the LICENSE file.
 
-package hdkeychain_test
+package hdkeychain
 
 import (
 	"testing"
-
-	"github.com/decred/dcrd/hdkeychain"
 )
 
 // bip0032MasterPriv1 is the master private extended key from the first set of
@@ -20,14 +18,14 @@ const bip0032MasterPriv1 = "dprv3hCznBesA6jBtmoyVFPfyMSZ1qYZ3WdjdebquvkEfmRfx" +
 // child from a master private extended key.
 func BenchmarkDeriveHardened(b *testing.B) {
 	b.StopTimer()
-	masterKey, err := hdkeychain.NewKeyFromString(bip0032MasterPriv1)
+	masterKey, err := NewKeyFromString(bip0032MasterPriv1)
 	if err != nil {
 		b.Errorf("Failed to decode master seed: %v", err)
 	}
 	b.StartTimer()
 
 	for i := 0; i < b.N; i++ {
-		masterKey.Child(hdkeychain.HardenedKeyStart)
+		masterKey.Child(HardenedKeyStart)
 	}
 }
 
@@ -35,7 +33,7 @@ func BenchmarkDeriveHardened(b *testing.B) {
 // (non-hardened) child from a master private extended key.
 func BenchmarkDeriveNormal(b *testing.B) {
 	b.StopTimer()
-	masterKey, err := hdkeychain.NewKeyFromString(bip0032MasterPriv1)
+	masterKey, err := NewKeyFromString(bip0032MasterPriv1)
 	if err != nil {
 		b.Errorf("Failed to decode master seed: %v", err)
 	}
@@ -50,7 +48,7 @@ func BenchmarkDeriveNormal(b *testing.B) {
 // key to a public extended key.
 func BenchmarkPrivToPub(b *testing.B) {
 	b.StopTimer()
-	masterKey, err := hdkeychain.NewKeyFromString(bip0032MasterPriv1)
+	masterKey, err := NewKeyFromString(bip0032MasterPriv1)
 	if err != nil {
 		b.Errorf("Failed to decode master seed: %v", err)
 	}
@@ -65,7 +63,7 @@ func BenchmarkPrivToPub(b *testing.B) {
 // extended key.
 func BenchmarkDeserialize(b *testing.B) {
 	for i := 0; i < b.N; i++ {
-		hdkeychain.NewKeyFromString(bip0032MasterPriv1)
+		NewKeyFromString(bip0032MasterPriv1)
 	}
 }
 
@@ -73,7 +71,7 @@ func BenchmarkDeserialize(b *testing.B) {
 // extended key.
 func BenchmarkSerialize(b *testing.B) {
 	b.StopTimer()
-	masterKey, err := hdkeychain.NewKeyFromString(bip0032MasterPriv1)
+	masterKey, err := NewKeyFromString(bip0032MasterPriv1)
 	if err != nil {
 		b.Errorf("Failed to decode master seed: %v", err)
 	}