--- a/hdkeychain/bench_test.go
+++ b/hdkeychain/bench_test.go
@@ -7,6 +7,8 @@ package hdkeychain
 
 import (
 	"testing"
+
+	"github.com/decred/dcrd/chaincfg"
 )
 
 // bip0032MasterPriv1 is the master private extended key from the first set of
@@ -18,7 +20,8 @@ const bip0032MasterPriv1 = "dprv3hCznBesA6jBtmoyVFPfyMSZ1qYZ3WdjdebquvkEfmRfx" +
 // child from a master private extended key.
 func BenchmarkDeriveHardened(b *testing.B) {
 	b.StopTimer()
-	masterKey, err := NewKeyFromString(bip0032MasterPriv1)
+	masterKey, err := NewKeyFromString(bip0032MasterPriv1,
+		&chaincfg.MainNetParams)
 	if err != nil {
 		b.Errorf("Failed to decode master seed: %v", err)
 	}
@@ -33,7 +36,8 @@ func BenchmarkDeriveHardened(b *testing.B) {
 // (non-hardened) child from a master private extended key.
 func BenchmarkDeriveNormal(b *testing.B) {
 	b.StopTimer()
-	masterKey, err := NewKeyFromString(bip0032MasterPriv1)
+	masterKey, err := NewKeyFromString(bip0032MasterPriv1,
+		&chaincfg.MainNetParams)
 	if err != nil {
 		b.Errorf("Failed to decode master seed: %v", err)
 	}
@@ -48,7 +52,8 @@ func BenchmarkDeriveNormal(b *testing.B) {
 // key to a public extended key.
 func BenchmarkPrivToPub(b *testing.B) {
 	b.StopTimer()
-	masterKey, err := NewKeyFromString(bip0032MasterPriv1)
+	masterKey, err := NewKeyFromString(bip0032MasterPriv1,
+		&chaincfg.MainNetParams)
 	if err != nil {
 		b.Errorf("Failed to decode master seed: %v", err)
 	}
@@ -63,7 +68,7 @@ func BenchmarkPrivToPub(b *testing.B) {
 // extended key.
 func BenchmarkDeserialize(b *testing.B) {
 	for i := 0; i < b.N; i++ {
-		NewKeyFromString(bip0032MasterPriv1)
+		NewKeyFromString(bip0032MasterPriv1, &chaincfg.MainNetParams)
 	}
 }
 
@@ -71,7 +76,8 @@ func BenchmarkDeserialize(b *testing.B) {
 // extended key.
 func BenchmarkSerialize(b *testing.B) {
 	b.StopTimer()
-	masterKey, err := NewKeyFromString(bip0032MasterPriv1)
+	masterKey, err := NewKeyFromString(bip0032MasterPriv1,
+		&chaincfg.MainNetParams)
 	if err != nil {
 		b.Errorf("Failed to decode master seed: %v", err)
 	}