--- a/hdkeychain/go.mod
+++ b/hdkeychain/go.mod
@@ -1,14 +1,12 @@
-module github.com/decred/dcrd/hdkeychain
+module github.com/decred/dcrd/hdkeychain/v2
 
 go 1.11
 
 require (
-	github.com/davecgh/go-spew v1.1.1 // indirect
 	github.com/decred/base58 v1.0.0
-	github.com/decred/dcrd/chaincfg v1.3.0
+	github.com/decred/dcrd/chaincfg v1.4.0
 	github.com/decred/dcrd/chaincfg/chainhash v1.0.1
 	github.com/decred/dcrd/dcrec v0.0.0-20190130161649-59ed4247a1d5
-	github.com/decred/dcrd/dcrec/edwards v0.0.0-20190130161649-59ed4247a1d5 // indirect
 	github.com/decred/dcrd/dcrec/secp256k1 v1.0.1
 	github.com/decred/dcrd/dcrutil v1.2.0
 	golang.org/x/crypto v0.0.0-20190131182504-b8fe1690c613 // indirect