… docs/assets/module_hierarchy.svg
… hdkeychain/README.md
… hdkeychain/go.sum