@matheusd 2019-03-22
An idea: also make the `NetParams` return the `masterKey` to be used when generating a new extended key. This would allow the package to be used in networks other than bitcoin/decred.