@davecgh 2019-03-22
That seems like a good idea.  It does mean the parameter should first be added to `chaincfg` and also the appropriate method introduced there so that the various `chaincfg` parameters implicitly satisfy the interface.  Otherwise, it would require the callers to write conversion code to get it to satisfy the interface which I'd like to avoid.