@davecgh 2019-03-22
After considering it a bit more, I believe that factoring out the master key would actually break multicurrency wallets since the BIP defines the seed as a specific value.  BIP44 relies on that.