@davecgh 2019-03-22
After discussing this with @jrick and considering it a bit more, changing the master key would actually break multi-currency wallets since the BIP defines the seed as a specific value and thus BIP44 implicitly relies on it.