@davecgh 2019-03-25
Updated to account for renamed HD extended key accessor funcs in `chaincfg`.