@davecgh 2019-03-29
This is ready for final review and ready to be merged.